Add model relationships to each model.
This specifies the 1-to-Many, Many-to-1, and Many-to-Many relationships on existing models. As per Laravel, relationships to many are specified with plural method names. Relationships from many to one are specified by a singular naming of the one model. Example: Board has many Post and Post belongs to Board. The board() method on Post model is singular while on the posts() method on the Board model is plural.